frontend: Add the option to use a getter function in the SimpleTable
The SimpleTable was previously designed only for selectinng the fields
of the given data object. However, a more flexible way of getting the
values from the object is to have a "getter" function assigned to the
column.

This patch implements that, but the actual use of it will come in a
later commit.