frontend: Ensure most of the simple list views show the loading state
This patch changes the default value for the items in that are passed to
most list views so it is null instead of [], thus setting up the loading
state in the SimpleTable.

A couple of list views aren't included in these changes because they're
items are gathered in a more complex way, so that will have to be done
later.