frontend: Add support for nested items in the Sidebar
The sidebar so far supported just one single level of items, but this
makes the organization of items difficult after a certain number that
has already been reached. This patch thus adds the support for nested
items to the sidebar component.

The actual reorganization of the sidebar's items will be done in a later
commit.