frontend: Add support for non-route items in Sidebar
The Sidebar supported only items that were assigned to a route. This
meant that each item had to load a different view that was assigned to
it.

However, now that the Sidebar can support nested items, it is useful to
have items that simply contain others and are not assigned to any view.
In this case, selecting such an item should load the route for the first
of its sub-items.