frontend: Don't require exact URL matches in the sidebar
The sidebar items were only matching the URL and setting themselves
as selected when the matching was strictly the same (no path prefixes).
This meant that any components not directly accessible from the sidebar,
like all the details views, would end be rendered without any entry
selected in the sidebar. This was not a problem when reaching the
mentioned views through the links within the app, but a direct URL
access meant that the sidebar would be left without a selected item.

This patch fixes that with a temporary solution. A proper solution will
be implemented when we introduce the pluggable system.