frontend: Show the metadata before the rest in the details views
The idea of showing the resource's particular info before the metadata
was because it's arguably more important than the common/metadata info.
However, this made things challenging in terms of layout (when there
were little or no info beyond the metadata).

So this patch makes the metadata be shown as the first element in the
section grid.