frontend: Add makeStatusLabel function
We already have a good component for showing status labels, but this
requires giving the name of the desired status. Since most of the times
we use binary statuses (i.e. success/error), then we could use a simpler
way for generating such labels. This patch provides that by simply
requiring the text for the label, and the expected text for the success
case.