frontend: Restyle the events table
Showing the whole message for each event means that a lot of information
is displayed and many of the columns are squeezed too.
This patch avoids that by showing the event's message (which may have
long text) in a tooltip instead.

This approach will likely be changed to a slide-down view instead, but
for now we use this as it's a quick/good-enough solution.