frontend: Add Redux with a filter reducer
We will need to set up a way to filter the resources based on the
namespace and other criteria.
The best way to do that, and especially taking into account that keeping
some of the filters on between views is a good idea, is to use Redux.

So this patch adds the redux + react-redux dependencies, and sets up a
store with the reducers for the filter. In the future, the use will be
of course extended beyond the filter use-cases.