frontend: Add SectionFilterHeader component
This new component is the same as the SectionHeader but introduces two
common actions related to filtering resources. Those are the search, and
the namespaces filters.

It uses also a new Autocomplete component for the namespaces filter,
which is based in the Material-UI lab module, but this may change soon
(because we want to get more control beyond what this module gives us,
so this is just a quick solution).