frontend: Allow notifying for multiple actions concurrently
The recent changes to have generic cancellable actions and a notifier
for them worked well but allowed only one message a time.
That is not enough if, e.g. an object is being deleted while another
is being updated.

These changes allow multiple actions to be used/notified at the same
time by using an object/map of actions + a stack of notifications.

This patch introduces a dependency on notistack.