frontend: Add cluster object editor dialog
This dialog has a few features that are useful when editing
Kubernetes objects, like the documentation, undoing changes, etc.

Besides the needed dependencies for the editor, react-monaco-editor and
monaco-editor-webpack-plugin, this patch also introduces
the react-app-rewired which replaces the react-app original scripts in
order to allow us to configure things like webpack without ejecting
(losing the "magic config" done by react-app).

The actual use of it will come in later commits.