frontend: Increase the title's width in the SectionHeader
The title in the SectionHeader was getting shrunk due to how the grid
was built, and thus was wrapping its text even when visually there was
plenty of space for it.

These changes make sure that the title has half of the header's width
to expand, which should be enough.