frontend: Select sidebar items according to the router's spec
The router already determines the URL and component to render, so
this patch adds a new "sidebar" property to the route specification
allowing to also determine which sidebar item to choose (or whether
to hide the whole sidebar).

Note that the sidebar item can still be selected from any component
by the use of the useSidebarItem hook, but having it in the router
specs is more self-contained and doesn't become a requirement for
creating new components.