frontend: Assign the events' colors depending on its type
The events' color was only green when the reason was "Ready", which
meant that all other event reasons showed up in red, leading to a
perception that the events were related to failures.

These changes fix that by assigning the colors per type of event.
Particularly, if the event type is "Normal" then it gets assigned the
normal color (gray), otherwise it gets assigned the warning color
(orange).