frontend: Fix count in Pods' status chart
The Pods' status chart was contrasting the pods with status as "Ready"
with the total number of pods. This meant that if there were pods whose
status was "Succeeded", then they would be counted towards the total
only and give the idea that some pods were not-ready or failed.

This patch fixes that by including the "Ready" and "Succeeded" pods in
the "positive" count of the chart, and dropping the label "Ready" from
the legend as it is no longer exclusively showing that status.