frontend: Don't display invalid data in the ConditionsTable
The ConditionsTable component was considering that the lastUpdateTime
and the reason were always available, so when these fields weren't, a
NaN was displayed.

These changes prevent that by using a "-" for that value instead. This
is a quick fix, as in the future we should likely just remove the whole
column from the table.