frontend: Do not duplicate items in the workload's overview
The overview of the workload items lists a number of cluster objects.
If there was more than one call to the listing callback (i.e. new
objects to be listed), then it would add those to the list while
keeping the existing ones; meaning it could end up in either outdated
objects in the list and duplicated ones.

This patch fixes that by always setting only the objects being listed,
i.e. not keeping the existing ones from previous calls.