frontend: Create a ViewDialog (as a read-only EditorDialog)
We need a way to show a resource's definition without using the edit
function (as it may not be editable). For that purpose, these changes
make the EditorDialog behave as read-only if the onSave property is
passed as null, and add a new ViewDialog component just to conveniently
use the mentioned editor in that way.

The actual use will come in a future commit.