backend: Add multi-cluster support to the backend
These changes allow the backend to serve several clusters instead of
just one, and rearrange the code in several files so things are bit
better abstracted.

It still serves one cluster from the previously used endpoint
(/cluster/{api...}) as this is needed until multi-cluster support is
added to the frontend.