backend: Reorganize the directory structure
The backend and frontend should be self-contained in their repective
paths as much as possible. The frontend was already in a good state but
the backend had the go.{mod,sum} files outside of the backend folder.

These changes move the backend/server.go into a new backend/cmd/ folder,
the mentioned go module related files into the backend folder, and
update the Makefile to reflect those changes.