frontend: Ensure the getConfig is not cluster-prefixed
The getConfig method needs to be called directly on the backend server
without any cluster prefix, but since the request method is using the
current URL to determine whether to prefix the request's path with the
cluster, this was resulting in an error.

This path fixes the issue above but should be temporary until the
current cluster management is changed.