frontend: Use Paper around PageGrid rather than around each section
As per new design, instead of using each section in the PageGrid
component as a paper section, we should use the PageGrid as paper
itself. So in order to accomplish this with minimum effort ATM, this
patch adds the paper wrapper to the PageGrid and changes the stylesheet
so it removes the elevation of paper within it (when we do remove the
paper wrappers from all components, we should drop this rule).