frontend: Add a title property to the SectionBox
Most of the SectionBoxes are used together with a SectionHeader
component with a title in it. So these changes aim to reduce the
repeated code by allowing a "title" property in the SectionBox
directly: if the title is a string, it'll be used in a SectionHeader,
if it is another type of React node, then it'll be used directly.