frontend: Use a responsive layout in the SectionHeader
The SectionHeader was assigning half of its content for the title and
the actions, yet this is not the best use of the space as it may force
a wrap of the contents when there is unused space by one of the items.

Thus, these changes use a responsive approach to make sure each item
uses as much space as needed until wrapping.