frontend: Redesign the SectionFilterHeader
This component's filter items were fully visible whenever used, even if
no filter was set. So in order to make it visually more appealing, we
show a simple filter button that needs to be clicked to reveal all the
filter options.

If a namespace is already set, then the full filter items will also be
immediately shown (i.e. not requiring the button to be clicked).