frontend: Simplify the MainInfoSection's meadata display
The MainInfoSection was using a MetadataDisplay component to display the
metadata, and a second NameValueTable for any important extra
information for the respective object.

This patch simplifies that by adding and using a new "extraRows"
property to the MetadataDisplay component, so it appends any wanted rows
to the main metadata table. This will allow us to apply a style to the
whole table and simplify the document's structure as well.