frontend: Fetch the clusters conf also from the chooser button
When accessing a view that is not the chooser directly (i.e. a URL
that is not the root one), then the configuration was not being fetched
from the server, resulting in not having the available clusters info
and thus having the chooser button become just a label.

These changes fix that by providing a new hook for getting the
configuration and using it in the chooser button instead of only in the
chooser dialog.