frontend: Create a KubeObject base class
The KubeObject was a base interface for all the cluster objects
implementation, but in order to make the code more encapsulated, we now
want to have the KubeObject as a real class from which other objects can
be created.

This patch adds that class and renames the mentioned interface to
KubeObjectInterface (as well as its references).