frontend: Add and use the remaining Workload related classes
The KubeWorkload type was used for representing objects that fit the
workload "type", but there were many changes to sometimes distinguish
between what kind of workload we needed (in order to fetch the data from
the server). This patch simplifies that by introducing classes for those
workloads, allowing things to be more encapsulated/abstracted, making
the code easier to maintain.