frontend: Remove use of DeleteMe in K8s object components
Instead of having each Details view implement the deletion callback for
its related Kubernetes object, we should just use the now abstracted
"delete" method of the KubeObject class. This means also that we don't
have to keep adding the DeleteButton component to each details view but
should instead add it to the MainInfoSection.

In order to make this transition easier, we're removing the deletion
logic from the Details component in this patch (the use of the delete
button in the MainInfoSection will be added in a later commit).