frontend: Remove direct EditButton use from serviceaccount/Details
The mentioned component was using an EditButton and setting up the
handler function for it. We want users to be able to edit any objects
(as long as they have permissions) so we will add a more automated way
of providing this functionality. Thus, this patch removes the use of the
EditButton from the ServiceAccount's details view.