frontend: Add update/put methods to the KubeObject class
These new methods will be used to make it easier to apply new changes to
a cluster, but also in the context of updating a K8s object, so we can
call update with the new configuration on an instance of KubeObject.

This will allow us to e.g. implement the EditButton's functionality in
a much more encapsulated way.