frontend: Use a partial KubeObjectInterface in the EditorDialog
The EditorDialog is used for editing Kubernetes objects, so it was
expecting a KubeObjectInterface as the passed item parameter. However,
since we will soon allow users to use the dialog to create a new object
from scratch, none of the KubeObjectInterface fields should be mandatory
in this case, and this patch implements that change.