frontend: Add a title property to the EditorDialog
We will want to use the EditDialog for also creating new resources from
scratch, and thus the default generated title for the dialog may not be
enough. In order to make the title customizable by the caller, we add a
title property to the dialog; if the property is not used, it falls back
to the previous case.