frontend: Add an errorMessage property to the EditorDialog
The EditorDialog component has an error/warning area at the bottom that
it uses by to show errors like YAML malformations. Since the dialog is
used for actions like editing a resource, it would be nice if the caller
code could set up an error message or warning for the user to see (e.g.
when editing a resource fails).

So this patch allows to override the error message with the mentioned
new property.