frontend: Add an apply method to apiProxy
We want to be able to apply (push or put) changes to a cluster, so this
patch adds a new method "apply" for that.
It uses the previously defined K8s classes and thus, in order to match
the desired resource with a class, it also adds a new index file to the
k8s dir which exports an object mapping classes names with the actual
classes.

The actual use of this new method will come in a later commit.