frontend: Add a back link to the MainInfoSection
We should have a more direct way to guide users to the list view when in
a component's details view. This link should take into account the
back/history link (so we can go back to a paginated table directly), but
that's an improvement to be done later. So this patch simply uses the
KubeObjects' listRoute property in the new back link button it adds to
the mentioned section.