backend: Fix handling the /config route
The /config route was defined after the / route, so when running the
cluster and serving the client app by it, since the base URL for the
requests is the same as the web app, the /config route was never tried
because the router would instead check if the path was in the web app.

This patch sets up the handling of /config before the / path, so it
fixes the issue mentioned above.