frontend: Add and use a getter for the KubeObject class name
When the code has been minified, then the use of a class' built-in name
will not be what's expected during the (unminified) development. This
leads to errors not only about how the class' name is presented in the
UI but also in other logic that depends on it.

We could fix this by configuring the minifier to avoid touching
functions' names, but that would require probably changing a lot of the
structure of the create-react-app setup we use. So this patch adds a
normal getter to the KubeObject class instead; this also allows more
flexibility e.g. if we need to use a customized class name.