backend: Fix serving static files for unregistered routes
The backend was only redirecting the root route to the static server and
this meant that if a user accessed a frontend URL that was not the root
directly, a 404 would be returned.

In order to fix the issue, this patch adds a new Single Page Application
handler (taken from mux's docs) that either serves files directly (when
they exist) or redirects the routes to the frontend's handler.