frontend: Use the location's origin as the base HTTP url
The frontend was using the location's URL (i.e. location without the
hash and search parameters) as the backend's URL; but this meant that if
the app's entry URL (the one accessed directly by the user) was
different from the root one, then it'd become the backend's URL and thus
any API calls would not resolve.
This patch fixes this by using the location's origin instead (base
protocol + host).