frontend: Increase number needed for the MetadataDictGrid to collapse
The mentioned component was collapsing if there were more than 2
metadata values to be shown. That meant that most uses of this component
would require the user to click the extend button in order to see more
metadata and that's not great UX. This patch simply increases the number
of items shown by default, but in the future we should be able to
configure it if needed.