frontend: Handle error for getMetrics
Currently our code assumes that we have the metric server installed
but this is not a requirement for running headlamp so at all places
where we use metrics in our code we should handle getMetrics errors.
This patch handles getMetrics error and for now adds a plain string
for error mentioning metric server is not installed.