desktop: Customize the app menu
We want to remove some of the properties from the app menu which are of
no use for us such as the Zoom property and we also at the same time
want to link to headlamps official github repo from the menu. Thus this
patch uses the template which is customizable and adds the required
rules to it.