frontend: Require user to manually refresh the SimpleTable's shown data
When the data shown in a SimpleTable was changed, then it would simply
show the new data and reset the page to 0. This represents a UX issue,
as the user may be looking for information in a table that suddenly
changes without warning.

To fix this, the mentioned component now shows a refresh button for
manually showing the new data, thus allowing users to keep reading the
contents they're reading or to load new ones.