frontend: Use a null events list by default in the cluster overview
Before the events list is loaded, we use its null state to show a loader
to the user in the EventsSection component. However, the data given to
it was set to an empty list by default, so the "No data was found"
message was being shown while loading the events.

This patch fixes that by simply allowing the events data to be null.