frontend: Fix the resource request endpoint for production
When running in production we were using the env
REACT_APP_BACKEND_SERVICES_URL but this makes the deployment process
more complex as we have to update this flag everytime we deploy to a
different domain (and thus build the frontend with that). This patch
updates the production request url to point to the current requesting
domain (accessed from window object).