frontend: Reset docs when the editor contents change
The docs' view is supposed to react to the contents of the editor. So
when the editor has a certain object kind and api version, the docs
should show the documentation for that object.
However, if there was an error shown (because e.g. the docs were not
found, or the contents were empty) then even if the contents were
corrected by the user, the error would remain.

To fix the mentioned issue, this patch resets the docs' error info
whenever the contents are changed.