frontend: Handle no param edge case for generating route path
We loop over the routes and render them accordingly. If a route requires
authentication and we don't have the required token we should be
redirected to the choser page but this was failing because while looping
and generating the route url if we pass an empty param with the cluster
url the generatePath(which is a helper from react-router-dom) will fail
because we are passing an empty param to it, In our usecase the empty
param actually means that its a request for the '/' choser page and not
the '/' cluster page. Thus this patch handles this usecase and returns
the choser route url in case of empty param.