frontend: Don't use current path as the homepage
We don't want to build the web app with a relative path which we are
currently doing by specifying the homepage property in package.json
for frontend dir. Instead we only need the relative path build for the
desktop app thus while building the desktop app we specify manually the
host path with the PUBLIC_URL env var.