frontend: Fix showing the full project logo's in the chooser dialog
The chooser dialog has the project logo on top but due to its styling
rules it was not showing the full width of the it depending on the
aspect ration of the logo used (i.e. the width would be cut out).

This patch increases the logo height and fixes the mentioned bug by
setting the width to 100%.