frontend: Allow to build with dev / local plugins
We removed the possibility of local plugins (modules that use the plugin
library but are built together with the frontend) since we now support
external plugins. However, local plugins may be needed when building
Headlamp with extra functionality, and also for making it easy to
develop plugins before "moving" them outside the Headlamp's tree.

So this patch fetches any external plugins, loads any local ones, and
only then it initializes the full plugins list.