frontend: Fix assigning header actions in Redux
The header actions are assigned using a name but the key used for the
name had not been updated in the UI reducer. As a result, the header
actions were always assigned to "undefined" and that meant not only that
we would lose the name, but also that only one action would be set up.
This patch corrects the key used for assigning the name.