Add plugins NPM pkg
Now that we support external plugins, the infrastructure to compile
those plugins was still not easy to configure. In order to make that
easier and also to provide other useful modules (like types, libs,
etc.) related to the plugins, this patch creates a new NPM package that
developers can use for building the Headlamp plugins externally.