frontend: Make class names in k8s.resourceClasses consistent
The class names have not been set in a consistent way: some are using
a capital first letter, and as a result the names in
k8s.resourceClasses were not the expected ones. We should make the
class names consistent but that has implications in the routes and
the sidebar, so for now we just capitalize the first letter when
adding the classes to k8s.resourceClasses. This is a temporary
measure until we have made de class names consistent.