frontend: Add correct list-route to the Ingress class
Since the list route name for this component's list view has an
irregular plural, it can not be computed the same way and this
resulted in a broke back button link when in the ingress details view.

This patch fixes that by simply overriding the listRoute property of
the Ingress class.