docker: Include a plugins dir in the image if it exists
We should include any plugins we want to ship in the image. So, in order
to do that, this patch copies anything in a local plugins folder into
the image and configures the Kubernetes deployment to use this plugins
dir. If the folder doesn't exist, then no plugins will be loaded by the
backend anyway.

In the future we should validate whether the plugins dir has contents
that resemble the plugins (plugins/*/main.js).