backend: Load contexts' info instead of clusters and use client auth
The backend was loading the clusters' info instead of the contexts which
in most cases is fine (as they correspond), but for other cases where a
different user is setup for the same cluster this was not supported.

So, when run with a KubeConfig, this patch loads the information from
the contexts defined there instead of the clusters. This means that we
now take the name from contexts and thus allow the same cluster to be
used twice if defined for different contexts.

The main motivation for this change however, was so that we could use
any client certificates associated with the contexts for accessing the
clusters without needing a token; this is the case for minikube for
example, for which we were requiring the user to set up a token before.
(Changes in the frontend are needed however, so the token requirement
isn't always prompted)

For the in-cluster config, a simple context is created just with the
in-cluster info.