frontend: Only require a token if the auth test fails without it
Now we support client certificates, which means that the user may not
have to set up a token to authenticate against a cluster. In order to
understand whether the user needs a token, this patch does a quick
auth test in the login dialog, and if it fails then a token will be
required, otherwise the user is redirected to correct view.