frontend: Add an error message generator method to the KubeObject class
Now that the KubeObject has a way to get either a list of the related
resources or an error, it's also very convenient to have a way to get
a message related to that error. Thus, a getErrorMessage is added to
the class so that its implementations can change the message to give
a bit of a better context in the future (if needed).