frontend: Check for errors when listing the objects in simple list views
If the user doesn't have the permissions to list some objects, then we
should reflect that in the UI. This patch does that by replacing the
use of KubeObject.useApiList by KubeObject.useList and using its error
to create an error message when needed.

The RoleBinding and Role list views are a bit more complex, so they will
be adapted in different commits.