frontend: Assign an interface to the makeKubeObject return
We need to generate the type declarations for some objects in order to
distribute them in the plugin-pkg module so plugin development can be
made easier by the use of types.

However, the way the actual KubeObject class is being created (as a
class defined in a method's body) makes the direct type inference
impossible. Thus, in order to fix that, this patch creates an
interface that partially matches the KubeObject class implementation
and uses it as the makeKubeObject return type, helping the type
inference.