frontend: Fix loader for role and roleBinding list view
We show a loader in table if the data is null but in case of role and
roleBinding list view we were initiating the data with empty array which
was leading to not show the loader when data was fetched.

This patch fixes the loader visibility for both mentioned views.