frontend: Move type declarations from reducers to lib modules
Some type declarations were being used by the reducers to declare the
states, and were also defined there. However, in the attempt to make
sure we can produce types declaration files that can be exported and
shipped in the plugins-pkg, we should move those types into lib related
modules (this way they're in a more logical place for existence, and
can be exported without referring the reducers).