plugins-pkg: Add basic types declaration files
We should allow plugin developers to use types related to plugins when
then develop them. The whole external plugin development story is still
being defined though, so for now we just export the types for the plugin
registry, Plugin class, and the router types they depend on.

Hopefully in the short term we will be able to provide a more solid and
defined package for defining plugins.