frontend: Reset error when getting results on useList / useMetrics
These hooks are supposed to make it very easy to use a results/error
associated with the resources we want, but while the error does reset
the list (sets it to null), the opposite is not true and this led to
an issue of not showing results after an error occurs (even if the
error is no longer happening).

This patch resets the error in the context explained above, which
fixes this issue.