frontend: Modify Tabs component
We need to use Tabs component when the sidebar is in shrink mode. This
also means we need to configure our Tabs component such that it can
take in default Values from outside and also reset to default whenever
asked for which should also be controlled from outside via prop.