frontend: Add new actions, action generators and modify reducers
We need two actions for sidebar shrink status and whether the sidebar
icon is clicked, Also we need to update the reducers accordingly.

Thus this patch adds two more properties to ui sidebar state about
sidebar shrink status and whether the sidebar icon is clicked and also
adds the relevant actions for them.