frontend: Handle invalid url scheme or protocols passed to WebSocket
We are using WebSocket in pod details view when establishing the
connection to the cluster terminal. But whenever a wrong url scheme or
protocol is passed to WebSocket constructor it breaks the entire app.

Thus this patch wraps the socket initialization in a try catch block and
any error that occurred while initializing the socket is handled by the
catch block.