frontend: Refactor App component
We want to use the theme when setting styles for the AppBar but right
now we were using useStyles outside the ThemeProvider which was leading
to use the default theme configs to be used in AppBar and not the one
we want to set.

This patch refactors the App component and segregates the entire App in
a seperate AppContainer component which receives the current theme to be
used via ThemeProvider.