backend: Don't stop the server if no config is found
We were using log.Fatal whenever there is no config present and as
log.Fatal internally gives a call to os.exit and because of which our
server stops. We don't want this behaviour and whenever there is no
config present we just want to log that info without stopping the
server.