ci: Upgrade go version and downgrade version for actions/setup-go
Workflow for backend was broken because in a latest update on github
set-env and add-path commands were deprecated and will no longer be
allowed and the version of actions/setup-go we uses these commands.

Thus this patch updates the version of go as well as specifies the
correct version for actions/setup-go.