frontend: Show an error message if there are no clusters configured
When the list of clusters is empty because none is configured, then the
UI was showing an endless spinner and a message saying it was waiting
for clusters.

This patch fixes that by initializing the clusters list as null and only
showing the "waiting" state until we get a clusters list result from the
server. If the list is empty, then an error message is shown.