backend: make it possible to serve SSL
This adds 4 new optional command line parameters:

  -ssl-cert filename
  -ssl-cert-key filename
  -ssl-cert-ca filename
  -ssl-client-cert-required

To serve SSL, the first two parameters are needed, without them HTTP is
served (like before).

The parameter -ssl-cert-ca might be used if the given SSL-certificate
uses a root CA which is not known by default.

The parameter -ssl-client-cert-required is by default false but might
be used to require that clients are having a valid client certificate.

So in order to serve SSL without something else which does SSL
termination, just mount the certificate(s) into the container (e.g. by
using a secret) and add the parameters -ssl-cert and -ssl-cert-key to
the args in the yaml with the deployment.