backend: Fix how we get home directory for preparing kubeconfig path
On platform like windows kubeconfig is generally created under the
current user directory thus the current way of fetching config doesn't
works on windows because we are always looking for the config in the
home directory which is wrong instead the correct location in which the
config is in the current users directory.

Thus this patch updates how we get HOME directory when preparing the
kubeconfig default path.