frontend: Fix race conditions when going back/forth in the AuthChooser
When choosing a cluster to authenticate to, if the chosen cluster was
unreachable and the user went back to the previous dialog, then when the
authentication attempt call was finished, it would could still redirect
the user when this component was already destroyed.

This patch cancels that effect if the user leaves the view.