frontend: Allow to set the back button link in common/MainInfoSection
The mentioned component will show a section with a back link which by
default will allow the user to go back to a resource's list view.

However, there may be cases where we want to link back to a different
view, e.g. when going from the workloads list into a workload type
resource's details view.

This patch maintains the default behavior for the link, but overrides it
if the backLink property is passed.