makefile: Update backend target to produce server.exe when on windows
When we are on windows platform we want to generate an executable for
our server binary. This is also important for our desktop app to run the
server because a plain binary file is not executable on windows.
Thus this patch updates the backend target to produce a server.exe when
the OS is windows.
For the desktop app now that we have a server.exe when on windows we
want to ship that as well into our electron packaging and thus we need
to create an entry for that in the extraResources property in
package.json. Also we don't want to commit the server executable so this
patch adds that as well to gitignore.