desktop: Setup for headless mode
We want to run the desktop app in headless mode this commit does the
intial setup required around it that includes accepting a headless mode
flag,bundling the frontend build directory with the desktop app and
if in headless mode running the server with that build directory passed
to the server.
For args parsing we also add yargs as a dependency so that users can
run headlamp with the help flag and see beautiful messages about
different flags that we have.