docs: Add docs for headless mode and windows installation guide
We now have headless mode in our app where users can run the desktop app
in there default browser but no mention of it is in the docs also we
miss the windows installation docs.
This patch adds docs on how headless mode can be used and also adds win
installation docs.