desktop: Port app code to typescript
This patch updates the main.js code that was used for desktop app to
typescript. We also add a new script that compiles the typescript file
and stores the resultant output file in electron/src/main.js.
We also don't want to commit the generated main.js file as it will be
auto generated on project build.