backend: Allow to configure the OIDC scopes
Apparently some OIDC providers don't ignore scopes that are passed but
are not available or allowed to be checked. This means that Headlamp
could not be use OIDC with some providers.

In order to fix that we make the OIDC scopes an option that needs to be
configured just like the client-id, secret, etc.