backend: Change -plugin-dir to use different defaults
These are the folders we use for the default plugin-dir:

  - the passed in pluginDir if it's not empty.
  - "./.plugins" if it exists.
  - ~/.config/Headlamp/plugins exists or it can be made
  - "./.plugins" if the ~/.config/Headlamp/plugins can't be made.

The user data config path is platform specific, so is different
on Windows for example.

For backwards compatibility, the Dockerfile now copies plugins
from the "./plugins" folder if they match the pattern
plugins/MyFolder/main.js. There is a new plugins folder coming to
the repo, so the .plugins folder is a new place to store built
plugins to be used for deployment.