Dockerfile: Make copying the plugins more robust
Copying the plugins was failing if the plugins folder was empty, because
in the Dockerfile we tried to copy from /headlamp/plugins but we first
had moved any plugins folder to plugins-old and if the folder was empty,
then the copy command wouldn't create the final plugins directory again.

Fixing that just required creating the mentioned plugins folder after
the move, but there was also the problem that the copy command was
copying everything in each plugins folder and not just the main.js file.