frontend: Add react-i18next for i18n
Use lazy loading of languages in the src/i18n/locales folder.

Integration with the Material UI Localization, so the localization
for those components works too.

There is some integration with the Monaco editor so that it can
show its messages in the selected language.

Number formatting, and date formatting can be done in an i18n way.