frontend: Use browser/OS light/dark theme preference
If the user makes a preference by selecting the theme from the UI,
then it uses that preference instead. Once the user preference is
selected it always uses that preference from then on.

The localStorage key was prefixed with headlamp to try and avoid
nameclashes. The older key was ignored, since it was set without
a user selecting a preference. This way the browser/OS preference
is used until they change it manually.

Also make the theme preference change reactive. So it changes
when the os/browser preference changes.