frontend: Allow to set the SimpleTable's sort function as a getter
This patch enables the "sort" attribute for a SimpleTable's column to be
a getter function, so this function is used for the default comparison
logic instead of the column's normal getter.

This will simplify how the data gets sorted when the normal getter
function cannot be used (e.g. when it returns a component).