frontend: Fix SideBar handling on reload/resize, and cleanup redux
Now loading and saving side bar open with localStorage
is done inside redux.

Also there is the concept of user interaction based opening.
This is useful for when the window is resized
from large(desktop) to xs(mobile).
With this resizing from large to xs keeps the sidebar state
good for the different window sizes unless the user has manually
toggled the sidebar. Otherwise going from large to xs size means
the side bar will be in the open state, whereas we want it in
the closed state on xs by default.

Also the preference for sidebar open or not is only stored when
it is in selected by the user via the button not via resizing.

Also cleaned up the topbar, layout and app container so they are
not passing down sidebar state.